\docType{package}
\name{bfork}
\alias{bfork-package}
\title{bfork: a paper thin wrapper of fork/waitpid for R}
\description{
    bfork provides a simple method for spawning and managing multiple processes.
}
\keyword{ fork }
\seealso{
    \code{\link{fork}}.\cr
    \code{\link{wait}}.\cr
    \code{\link{waitall}}.
}
